Refactor model helpers into own file to allow helper import when apps not loaded yet
