+++ InvenTree/InvenTree/serializers.py
@@ -24,7 +24,7 @@ from taggit.serializers import TaggitSerializer
 from common.models import InvenTreeSetting
 from common.settings import currency_code_default, currency_code_mappings
 from InvenTree.fields import InvenTreeRestURLField, InvenTreeURLField
-from InvenTree.helpers import download_image_from_url
+from InvenTree.helpers_model import download_image_from_url
 
 
 class InvenTreeMoneySerializer(MoneyField):
