+++ InvenTree/common/tasks.py
@@ -11,7 +11,7 @@ from django.utils import timezone
 
 import feedparser
 
-from InvenTree.helpers import getModelsWithMixin
+from InvenTree.helpers_model import getModelsWithMixin
 from InvenTree.models import InvenTreeNotesMixin
 from InvenTree.tasks import ScheduledTask, scheduled_task
